refactor : domain/service 에 따라 클래스 분리
역할과 책임에 따라 패키지 구분 후 클래스 분리
