+++ src/test/java/lotto/LottoResultTest.java
@@ -5,6 +5,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import lotto.domain.Lotto;
+import lotto.domain.Ranking;
+import lotto.domain.WinningNumbers;
+import lotto.service.Calculator;
+import lotto.service.LottoResult;
 import org.junit.jupiter.api.Test;
 
 public class LottoResultTest {
