+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,8 @@
 package lotto;
 
 import java.util.Arrays;
+import lotto.domain.Lotto;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
